Moved config.js into src/conf folder
